:recycle: 移除 sv 模式工具栏高亮
